Extract board thread read models
